@@ -1,4 +1,4 @@
-<style>
+<style lang="scss">
   .side-nav {
     width: 100%;
     box-sizing: border-box;
@@ -131,11 +131,6 @@
             </a>
           </li>
           <li class="nav-item">
-            <a class="sponsor" href="https://www.duotai.net/?utm_source=element" target="_blank">
-              <img src="~examples/assets/images/duotai.svg" alt="duotai">
-            </a>
-          </li>
-          <li class="nav-item">
             <a class="sponsor" href="https://www.duohui.cn/?utm_source=element&utm_medium=web&utm_campaign=element-index" target="_blank">
               <img src="~examples/assets/images/duohui.svg" alt="duohui">
             </a>
@@ -180,7 +175,7 @@
               <li
                 class="nav-item"
                 v-for="(navItem, key) in group.list"
-                v-if="!navItem.disabled"
+                v-show="!navItem.disabled"
                 :key="key">
                 <router-link
                   active-class="active"
