@@ -2,14 +2,19 @@ import Vue from 'vue';
 import entry from './app';
 import VueRouter from 'vue-router';
 import Element from 'main/index.js';
-import 'packages/theme-chalk/src/index.scss';
+import hljs from 'highlight.js';
 import routes from './route.config';
-import demoBlock from './components/demo-block.vue';
-import MainFooter from './components/footer.vue';
-import MainHeader from './components/header.vue';
+import demoBlock from './components/demo-block';
+import MainFooter from './components/footer';
+import MainHeader from './components/header';
 import SideNav from './components/side-nav';
 import FooterNav from './components/footer-nav';
-import title from './i18n/title.json';
+import title from './i18n/title';
+
+import 'packages/theme-chalk/src/index.scss';
+import './demo-styles/index.scss';
+import './assets/styles/common.css';
+import './assets/styles/fonts/style.css';
 
 Vue.use(Element);
 Vue.use(VueRouter);
@@ -26,6 +31,11 @@ const router = new VueRouter({
 });
 
 router.afterEach(route => {
+  // https://github.com/highlightjs/highlight.js/issues/909#issuecomment-131686186
+  Vue.nextTick(() => {
+    const blocks = document.querySelectorAll('pre code:not(.hljs)');
+    Array.prototype.forEach.call(blocks, hljs.highlightBlock);
+  });
   const data = title[route.meta.lang];
   for (let val in data) {
     if (new RegExp('^' + val, 'g').test(route.name)) {
@@ -34,9 +44,10 @@ router.afterEach(route => {
     }
   }
   document.title = 'Element';
+  ga('send', 'event', 'PageView', route.name);
 });
 
 new Vue({ // eslint-disable-line
-  render: h => h(entry),
+  ...entry,
   router
 }).$mount('#app');
