@@ -11,9 +11,8 @@
         <span class="el-alert__title" :class="[ isBoldTitle ]" v-if="title || $slots.title">
           <slot name="title">{{ title }}</slot>
         </span>
-        <slot>
-          <p class="el-alert__description" v-if="description">{{ description }}</p>
-        </slot>
+        <p class="el-alert__description" v-if="$slots.default && !description"><slot></slot></p>
+        <p class="el-alert__description" v-if="description && !$slots.default">{{ description }}</p>
         <i class="el-alert__closebtn" :class="{ 'is-customed': closeText !== '', 'el-icon-close': closeText === '' }" v-show="closable" @click="close()">{{closeText}}</i>
       </div>
     </div>
