@@ -1,5 +1,5 @@
 import { getCell, getColumnByCell, getRowIdentity, getToolTip } from './util';
-import { getStyle, hasClass, addClass, removeClass } from 'element-ui/src/utils/dom';
+import { getStyle, hasClass } from 'element-ui/src/utils/dom';
 import ElCheckbox from 'element-ui/packages/checkbox';
 
 import LayoutObserver from './layout-observer';
@@ -27,6 +27,29 @@ export default {
 
   render(h) {
     const columnsHidden = this.columns.map((column, index) => this.isColumnHidden(index));
+    let rows = this.data;
+    if (this.store.states.lazy && Object.keys(this.store.states.lazyTreeNodeMap).length) {
+      rows = rows.reduce((prev, item) => {
+        prev.push(item);
+        const rowKey = this.store.table.getRowKey(item);
+        const parent = this.store.states.treeData[rowKey];
+        if (parent && parent.children) {
+          const tmp = [];
+          const traverse = (children) => {
+            if (!children) return;
+            children.forEach(key => {
+              tmp.push(this.store.states.lazyTreeNodeMap[key]);
+              if (this.store.states.treeData[key]) {
+                traverse(this.store.states.treeData[key].children);
+              }
+            });
+          };
+          traverse(parent.children);
+          prev = prev.concat(tmp);
+        }
+        return prev;
+      }, []);
+    }
     return (
       <table
         class="el-table__body"
@@ -40,22 +63,46 @@ export default {
         </colgroup>
         <tbody>
           {
-            this._l(this.data, (row, $index) =>
-              [<tr
+            this._l(rows, (row, $index) => {
+              const rowKey = this.table.rowKey ? this.getKeyOfRow(row, $index) : $index;
+              const treeNode = this.treeData[rowKey];
+              const rowClasses = this.getRowClass(row, $index);
+              if (treeNode) {
+                rowClasses.push('el-table__row--level-' + treeNode.level);
+              }
+              const tr = (<tr
+                v-show={ treeNode ? treeNode.display : true }
                 style={ this.rowStyle ? this.getRowStyle(row, $index) : null }
-                key={ this.table.rowKey ? this.getKeyOfRow(row, $index) : $index }
+                key={ rowKey }
                 on-dblclick={ ($event) => this.handleDoubleClick($event, row) }
                 on-click={ ($event) => this.handleClick($event, row) }
                 on-contextmenu={ ($event) => this.handleContextMenu($event, row) }
                 on-mouseenter={ _ => this.handleMouseEnter($index) }
                 on-mouseleave={ _ => this.handleMouseLeave() }
-                class={ [this.getRowClass(row, $index)] }>
+                class={ rowClasses }>
                 {
                   this._l(this.columns, (column, cellIndex) => {
                     const { rowspan, colspan } = this.getSpan(row, column, $index, cellIndex);
                     if (!rowspan || !colspan) {
                       return '';
                     } else {
+                      const data = {
+                        store: this.store,
+                        _self: this.context || this.table.$vnode.context,
+                        row,
+                        column,
+                        $index
+                      };
+                      if (cellIndex === this.firstDefaultColumnIndex && treeNode) {
+                        data.treeNode = {
+                          hasChildren: treeNode.hasChildren || (treeNode.children && treeNode.children.length),
+                          expanded: treeNode.expanded,
+                          indent: treeNode.level * this.treeIndent,
+                          level: treeNode.level,
+                          loaded: treeNode.loaded,
+                          rowKey
+                        };
+                      }
                       return (
                         <td
                           style={ this.getCellStyle($index, cellIndex, row, column) }
@@ -68,13 +115,7 @@ export default {
                             column.renderCell.call(
                               this._renderProxy,
                               h,
-                              {
-                                row,
-                                column,
-                                $index,
-                                store: this.store,
-                                _self: this.context || this.table.$vnode.context
-                              },
+                              data,
                               columnsHidden[cellIndex]
                             )
                           }
@@ -83,56 +124,24 @@ export default {
                     }
                   })
                 }
-              </tr>,
-              this.store.isRowExpanded(row)
-                ? (<tr>
-                  <td colspan={ this.columns.length } class="el-table__expanded-cell">
-                    { this.table.renderExpanded ? this.table.renderExpanded(h, { row, $index, store: this.store }) : ''}
-                  </td>
-                </tr>)
-                : ''
-              ]
-            )
+              </tr>);
+              if (this.hasExpandColumn && this.store.isRowExpanded(row)) {
+                return [
+                  tr,
+                  <tr>
+                    <td colspan={ this.columns.length } class="el-table__expanded-cell">
+                      { this.table.renderExpanded ? this.table.renderExpanded(h, { row, $index, store: this.store }) : ''}
+                    </td>
+                  </tr>
+                ];
+              } else {
+                return tr;
+              }
+            })
           }
         </tbody>
       </table>
     );
-  },
-
-  watch: {
-    'store.states.hoverRow'(newVal, oldVal) {
-      if (!this.store.states.isComplex) return;
-      const el = this.$el;
-      if (!el) return;
-      const tr = el.querySelector('tbody').children;
-      const rows = [].filter.call(tr, row => hasClass(row, 'el-table__row'));
-      const oldRow = rows[oldVal];
-      const newRow = rows[newVal];
-      if (oldRow) {
-        removeClass(oldRow, 'hover-row');
-      }
-      if (newRow) {
-        addClass(newRow, 'hover-row');
-      }
-    },
-    'store.states.currentRow'(newVal, oldVal) {
-      if (!this.highlight) return;
-      const el = this.$el;
-      if (!el) return;
-      const data = this.store.states.data;
-      const tr = el.querySelector('tbody').children;
-      const rows = [].filter.call(tr, row => hasClass(row, 'el-table__row'));
-      const oldRow = rows[data.indexOf(oldVal)];
-      const newRow = rows[data.indexOf(newVal)];
-      if (oldRow) {
-        removeClass(oldRow, 'current-row');
-      } else {
-        [].forEach.call(rows, row => removeClass(row, 'current-row'));
-      }
-      if (newRow) {
-        addClass(newRow, 'current-row');
-      }
-    }
   },
 
   computed: {
@@ -142,6 +151,10 @@ export default {
 
     data() {
       return this.store.states.data;
+    },
+
+    treeData() {
+      return this.store.states.treeData;
     },
 
     columnsCount() {
@@ -166,6 +179,23 @@ export default {
 
     columns() {
       return this.store.states.columns;
+    },
+
+    hasExpandColumn() {
+      return this.columns.some(({ type }) => type === 'expand');
+    },
+
+    firstDefaultColumnIndex() {
+      for (let index = 0; index < this.columns.length; index++) {
+        if (this.columns[index].type === 'default') {
+          return index;
+        }
+      }
+      return 0;
+    },
+
+    treeIndent() {
+      return this.store.states.indent;
     }
   },
 
@@ -239,6 +269,10 @@ export default {
         classes.push('current-row');
       }
 
+      if (rowIndex === this.store.states.hoverRow) {
+        classes.push('hover-row');
+      }
+
       if (this.stripe && rowIndex % 2 === 1) {
         classes.push('el-table__row--striped');
       }
@@ -256,7 +290,7 @@ export default {
         classes.push('expanded');
       }
 
-      return classes.join(' ');
+      return classes;
     },
 
     getCellStyle(rowIndex, columnIndex, row, column) {
@@ -370,7 +404,7 @@ export default {
           table.$emit(`cell-${name}`, row, column, cell, event);
         }
       }
-      table.$emit(`row-${name}`, row, event, column);
+      table.$emit(`row-${name}`, row, column, event);
     },
 
     handleExpandClick(row, e) {
