@@ -224,6 +224,26 @@
   import TableBody from './table-body';
   import TableHeader from './table-header';
   import TableFooter from './table-footer';
+  import { getRowIdentity } from './util';
+
+  const flattenData = function(data) {
+    if (!data) return data;
+    let newData = [];
+    const flatten = arr => {
+      arr.forEach((item) => {
+        newData.push(item);
+        if (Array.isArray(item.children)) {
+          flatten(item.children);
+        }
+      });
+    };
+    flatten(data);
+    if (data.length === newData.length) {
+      return data;
+    } else {
+      return newData;
+    }
+  };
 
   let tableIdSeed = 1;
 
@@ -311,7 +331,16 @@
       selectOnIndeterminate: {
         type: Boolean,
         default: true
-      }
+      },
+
+      indent: {
+        type: Number,
+        default: 16
+      },
+
+      lazy: Boolean,
+
+      load: Function
     },
 
     components: {
@@ -451,6 +480,55 @@
 
       toggleAllSelection() {
         this.store.commit('toggleAllSelection');
+      },
+
+      getRowKey(row) {
+        const rowKey = getRowIdentity(row, this.store.states.rowKey);
+        if (!rowKey) {
+          throw new Error('if there\'s nested data, rowKey is required.');
+        }
+        return rowKey;
+      },
+
+      getTableTreeData(data) {
+        const treeData = {};
+        const traverse = (children, parentData, level) => {
+          children.forEach(item => {
+            const rowKey = this.getRowKey(item);
+            treeData[rowKey] = {
+              display: false,
+              level
+            };
+            parentData.children.push(rowKey);
+            if (Array.isArray(item.children) && item.children.length) {
+              treeData[rowKey].children = [];
+              treeData[rowKey].expanded = false;
+              traverse(item.children, treeData[rowKey], level + 1);
+            }
+          });
+        };
+        if (data) {
+          data.forEach(item => {
+            const containChildren = Array.isArray(item.children) && item.children.length;
+            if (!(containChildren || item.hasChildren)) return;
+            const rowKey = this.getRowKey(item);
+            const treeNode = {
+              level: 0,
+              expanded: false,
+              display: true,
+              children: []
+            };
+            if (containChildren) {
+              treeData[rowKey] = treeNode;
+              traverse(item.children, treeData[rowKey], 1);
+            } else if (item.hasChildren && this.lazy) {
+              treeNode.hasChildren = true;
+              treeNode.loaded = false;
+              treeData[rowKey] = treeNode;
+            }
+          });
+        }
+        return treeData;
       }
     },
 
@@ -582,6 +660,8 @@
       data: {
         immediate: true,
         handler(value) {
+          this.store.states.treeData = this.getTableTreeData(value);
+          value = flattenData(value);
           this.store.commit('setData', value);
           if (this.$ready) {
             this.$nextTick(() => {
@@ -633,7 +713,9 @@
       const store = new TableStore(this, {
         rowKey: this.rowKey,
         defaultExpandAll: this.defaultExpandAll,
-        selectOnIndeterminate: this.selectOnIndeterminate
+        selectOnIndeterminate: this.selectOnIndeterminate,
+        indent: this.indent,
+        lazy: this.lazy
       });
       const layout = new TableLayout({
         store,
